feat: arch example
