+++ .env.example
@@ -1,0 +1,2 @@
+NEXT_PUBLIC_SIEHP_URL=http://localhost:3000
+NEXT_PUBLIC_SIEHP_API_URL=http://localhost:8080
