+++ src/app/example/layout.tsx
@@ -1,4 +1,4 @@
-import MakeExampleLayout from "@/main/factories/layouts/MakeExampleLayout";
+import MakeExampleLayout from "@/shared/infra/factories/layouts/MakeExampleLayout";
 import { ReactNode } from "react";
 
 export default function ExampleLayout({ children }: { children: ReactNode }) {
