+++ src/app/example/page.tsx
@@ -1,4 +1,4 @@
-import MakeExamplePage from "@/main/factories/pages/MakeExamplePage";
+import MakeExamplePage from "@/shared/infra/factories/pages/MakeExamplePage";
 
 export default function ExamplePage() {
   return <MakeExamplePage />;
