--- src/data/usecases/index.ts
@@ -1,1 +1,0 @@
-export * from './remote-example'
