--- src/main/adapters/example-adapter.ts
@@ -1,8 +1,0 @@
-import { ExampleAdapterDTO, GetExampleResponseDTO } from "@/domain/dtos";
-
-export function exampleAdapter({error, example}: GetExampleResponseDTO): ExampleAdapterDTO {
-    console.log(error)
-    return {
-        example: example,
-    }
-}
