--- src/main/factories/pages/MakeExamplePage/index.tsx
@@ -1,10 +1,0 @@
-import ExamplePage from '@/presentation/pages/ExamplePage'
-
-const MakeExamplePage = () => {
-  return (
-    // Pode passar providers aqui tbm se precisar
-    <ExamplePage/>
-  )
-}
-
-export default MakeExamplePage
