+++ src/shared/config/libs/cypress/.gitignore
