+++ src/shared/domain/dtos/get-example-dto.ts
