+++ src/shared/domain/errors/example-error.ts
