+++ src/shared/domain/errors/index.ts
