+++ src/shared/domain/errors/out-of-context-error.ts
