+++ src/shared/domain/models/example-model.ts
