+++ src/shared/domain/usecases/example.ts
