+++ src/shared/infra/factories/layouts/MakeExampleLayout/types.ts
