+++ src/shared/infra/factories/pages/MakeExamplePage/index.tsx
@@ -1,0 +1,11 @@
+import ExamplePage from "@/app/example/page"
+
+
+const MakeExamplePage = () => {
+  return (
+    // Pode passar providers aqui tbm se precisar
+    <ExamplePage/>
+  )
+}
+
+export default MakeExamplePage
