+++ src/shared/infra/factories/pages/MakeExamplePage/tests/fixtures/example-fixture.json
@@ -1,0 +1,3 @@
+{
+    "data": "example data"
+}
