+++ src/shared/infra/presentation/components/ExampleComponent/index.tsx
