+++ src/shared/infra/presentation/components/ExampleComponent/types.ts
