+++ src/shared/infra/presentation/layouts/ExampleLayout/types.ts
