+++ src/shared/infra/services/data/adapters/example-adapter.ts
@@ -1,0 +1,8 @@
+import { ExampleAdapterDTO, GetExampleResponseDTO } from "@/shared/domain/dtos"
+
+export function exampleAdapter({error, example}: GetExampleResponseDTO): ExampleAdapterDTO {
+    console.log(error)
+    return {
+        example: example,
+    }
+}
