+++ src/shared/infra/services/data/adapters/index.ts
