+++ src/shared/infra/services/data/usecases/index.ts
@@ -1,0 +1,1 @@
+export * from './remote-example-usecase'
