+++ src/shared/infra/utils/constants/example-constant.ts
