+++ src/shared/infra/utils/constants/index.ts
@@ -1,0 +1,1 @@
+export * from './example-constant'
