+++ src/shared/infra/utils/functions/example-util-function.ts
