+++ src/shared/infra/utils/functions/index.ts
